% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{rip_info}
\alias{rip_info}
\title{Extract all info from a date or datetime}
\usage{
rip_info(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the information from.}
}
\value{
A named list of size \code{8} containing the information about \code{x}.
}
\description{
\code{rip_info()} extracts all information from a date or datetime object.
}
\examples{
x <- as.Date("2019-01-01")
rip_info(x)

# Identical dates, but in different years
a <- as.POSIXct("1969-12-31 19:00:00", tz = "America/New_York")
b <- as.POSIXct("1970-01-01 00:00:00", tz = "UTC")

identical(as.numeric(a), as.numeric(b))

rip_info(a)
rip_info(b)
}
