% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second.R
\name{rip_second}
\alias{rip_second}
\title{Extract the seconds from a date or datetime}
\usage{
rip_second(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the seconds from.}
}
\value{
A double vector the same size as \code{x} containing the seconds.
}
\description{
\code{rip_second()} extracts the seconds from a date or datetime object.
}
\examples{
x <- as.POSIXct("2019-01-01 00:00:05", "UTC")
rip_second(x)
}
