% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour.R
\name{time_hour}
\alias{time_hour}
\title{Extract the hour from a date or datetime}
\usage{
time_hour(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the hour from.}
}
\value{
An integer vector the same size as \code{x} containing the hours.
}
\description{
\code{time_hour()} extracts the hour from a date or datetime object.
}
\examples{
# Identical dates, but in different hours
x <- as.POSIXct("1969-12-31 19:00:00", tz = "America/New_York")
y <- as.POSIXct("1970-01-01 00:00:00", tz = "UTC")

identical(as.numeric(x), as.numeric(y))

time_hour(x)
time_hour(y)
}
