% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minute.R
\name{time_minute}
\alias{time_minute}
\title{Extract the minutes from a date or datetime}
\usage{
time_minute(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the minutes from.}
}
\value{
An integer vector the same size as \code{x} containing the minutes.
}
\description{
\code{time_minute()} extracts the minutes from a date or datetime object.
}
\examples{
x <- as.POSIXct("1969-12-31 19:05:00", tz = "America/New_York")
time_minute(x)
}
