% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{time_month}
\alias{time_month}
\title{Extract the month from a date or datetime}
\usage{
time_month(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the year from.}
}
\value{
An integer vector the same size as \code{x} containing the months. \code{1}
aligns with January, and \code{12} is December.
}
\description{
\code{time_month()} extracts the month from a date or datetime object.
}
\examples{
x <- as.Date("2019-01-01")
time_month(x)

# Identical dates, but in different months
a <- as.POSIXct("1969-12-31 19:00:00", tz = "America/New_York")
b <- as.POSIXct("1970-01-01 00:00:00", tz = "UTC")

identical(as.numeric(a), as.numeric(b))

time_month(a)
time_month(b)
}
