% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wday.R
\name{time_wday}
\alias{time_wday}
\title{Extract the day of the week from a date or datetime}
\usage{
time_wday(x)
}
\arguments{
\item{x}{\verb{<Date / POSIXct / POSIXlt>}

The vector to extract the day of the week from.}
}
\value{
An integer vector the same size as \code{x} containing the day of the
week values. Monday corresponds to \code{1}, and Sunday is \code{7}.
}
\description{
\code{time_mday()} extracts the day of the week from a date or datetime object.
}
\examples{
# Identical dates, but in different hours
x <- as.POSIXct("1969-12-31 19:00:00", tz = "America/New_York")
y <- as.POSIXct("1970-01-01 00:00:00", tz = "UTC")

identical(as.numeric(x), as.numeric(y))

time_wday(x)
time_wday(y)
}
