% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timerip-package.R
\docType{package}
\name{timerip-package}
\alias{timerip}
\alias{timerip-package}
\title{timerip: Extract Datetime Components}
\description{
Extract datetime components easily and efficiently. The R Core
    group is the original author of the code made available with slight
    modifications by this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/timerip}
  \item Report bugs at \url{https://github.com/DavisVaughan/timerip/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
