% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timewarp-package.R
\docType{package}
\name{timewarp-package}
\alias{timewarp}
\alias{timewarp-package}
\title{timewarp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
