% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timewarp-package.R
\docType{package}
\name{timewarp-package}
\alias{timewarp}
\alias{timewarp-package}
\title{timewarp: Extract Datetime Components}
\description{
Extract datetime components easily and efficiently. The R Core
    group is the original author of the code made available with slight
    modifications by this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/timewarp}
  \item Report bugs at \url{https://github.com/DavisVaughan/timewarp/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
